icosahedron planet done

+++ src2/la/vector_op.c
@@ -6,7 +6,7 @@
 /*   By: jaelee <jaelee@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/05/10 14:29:07 by jaelee            #+#    #+#             */
-/*   Updated: 2019/06/03 09:29:37 by jaelee           ###   ########.fr       */
+/*   Updated: 2019/06/06 22:38:09 by jaelee           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -57,7 +57,7 @@ void	vec3_normalize(float *vec)
 	float	normalize_factor;
 
 	normalize_factor =
-	1.f / sqrt((vec[0] * vec[0]) + (vec[1] * vec[1]) + (vec[2] * vec[2]));
+	1.f / sqrtf((vec[0] * vec[0]) + (vec[1] * vec[1]) + (vec[2] * vec[2]));
 	vec[0] = vec[0] * normalize_factor;
 	vec[1] = vec[1] * normalize_factor;
 	vec[2] = vec[2] * normalize_factor;
